fix crash with warped surfaces
there are weird assumptions all over the place, but
no idea why this is done that way.
R_PolygonScanLeftEdge can process less spans than
R_PolygonScanRightEdge which then works on
uninitialized data and outputs garbage, but this
somehow is necessary for these surfaces to render.
this might still be a point of failure.